Move types declared in index.tsx to their own files
