+++ ui/src/Common/Branding/Branding.test.tsx
@@ -17,8 +17,8 @@
 import React from 'react';
 import {MemoryRouter} from 'react-router-dom';
 import Branding from './Branding';
-import {RunConfig} from '../../index';
 import {render, screen} from '@testing-library/react';
+import {RunConfig} from 'Types/RunConfig';
 
 describe('Branding', () => {
     const expectedUrl = 'url-dot-com';
