+++ ui/src/Types/RunConfig.ts
@@ -1,0 +1,27 @@
+/*
+ * Copyright (c) 2022 Ford Motor Company
+ * All rights reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+export interface RunConfig {
+    auth_enabled: boolean;
+    ford_labs_url: string;
+    invite_users_to_space_enabled: boolean;
+    adfs_url_template: string;
+    adfs_client_id: string;
+    adfs_resource: string;
+    flagsmith_environment_id: string;
+    flagsmith_url: string;
+}
